+++ package.json
@@ -20,7 +20,14 @@
   "devDependencies": {
     "babel-core": "^6.17.0",
     "babel-loader": "^6.2.5",
+    "babel-preset-react": "^6.16.0",
     "babel-preset-es2015": "^6.16.0",
+    "css-loader": "^0.25.0",
+    "sass-loader": "^4.0.2",
+    "style-loader": "^0.13.1",
     "webpack": "^1.13.2"
+  },
+  "dependencies": {
+    "react": "^15.3.2"
   }
 }
