+++ src/ReactPopupTooltip.scss
@@ -1,0 +1,85 @@
+$backgroundColor:#999;
+$borderRadius:5px;
+.Popup-Tooltip-Container{
+
+  display: inline-flex;
+  flex-direction: column;
+  align-items: center;
+  position:absolute;
+
+  .Popup-Tooltip-Content {
+    min-width:100px;
+    //max-width:400px;
+    background-color:$backgroundColor;
+    padding:5px 10px;
+    border-radius:$borderRadius;
+    -webkit-box-shadow: 0px 0px 2px 0px rgba(0,0,0,1);
+    -moz-box-shadow: 0px 0px 2px 0px rgba(0,0,0,1);
+    box-shadow: 0px 0px 2px 0px rgba(0,0,0,1);
+  }
+  .Popup-Tooltip-Arrow {
+    width: 0;
+    height: 0;
+    border-style:solid;
+    z-index:2;
+    border-width: 0 7.5px 10px;
+    border-color:transparent transparent $backgroundColor transparent;
+  }
+  //default or vertical set
+  &[data-align='start'] {
+
+    .Popup-Tooltip-Arrow {
+      align-self:flex-start;
+      border-width: 10px 0 0 10px;
+      border-color: transparent transparent transparent $backgroundColor;
+    }
+    .Popup-Tooltip-Content {
+      border-top-left-radius: 0;
+    }
+
+  }
+  &[data-align='end'] {
+
+    .Popup-Tooltip-Arrow {
+      align-self:flex-end;
+      border-width: 0 0 10px 10px;
+      border-color: transparent transparent $backgroundColor transparent;
+    }
+    .Popup-Tooltip-Content {
+      border-top-right-radius: 0;
+    }
+
+  }
+  &[data-justify='end']{
+    .Popup-Tooltip-Content {
+      border-radius: $borderRadius;
+    }
+    .Popup-Tooltip-Arrow {
+      border-width: 10px 7.5px 0 7.5px;
+      border-color: $backgroundColor transparent transparent transparent;
+      order:1;
+    }
+    &[data-align='start'] {
+      .Popup-Tooltip-Content {
+        border-bottom-left-radius: 0;
+      }
+      .Popup-Tooltip-Arrow {
+        border-width: 10px 10px 0 0;
+        border-color: $backgroundColor transparent transparent transparent;
+      }
+    }
+    &[data-align='end'] {
+      .Popup-Tooltip-Content {
+        border-bottom-right-radius: 0;
+      }
+      .Popup-Tooltip-Arrow {
+        border-width: 0 10px 10px 0;
+        border-color: transparent $backgroundColor transparent transparent;
+      }
+    }
+  }
+
+  &.hidden {
+    display:none;
+  }
+}
