+++ webpack.config.js
@@ -16,9 +16,12 @@ module.exports = {
         loader: 'babel-loader',
         exclude: /node_modules/,
         query: {
-          presets: ['es2015']
+          presets: ['es2015', 'react']
         }
-      }
+      },
+      {
+        test:/\.scss$/,
+        loaders:['style','css','sass']},
     ]
   },
 };
